# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hospital Drug Inventory Management System",
    "description": "AI-powered drug inventory optimization using demand prediction and clustering",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/hospital/upload-stock": {
      "post": {
        "tags": [
          "data upload"
        ],
        "summary": "Upload Stock",
        "description": "Upload current stock data via CSV file (Hospital users only).\n\n**Authentication:** Requires hospital_admin or hospital_staff role\n\n**CSV Format:**\n```\nmedicine_id,medicine_name,medicine_quantity,medicine_expiry\nMED001,Paracetamol 500mg,5000,2025-12-31\n```\n\n**Returns:**\n- `message`: Success/error message\n- `records_processed`: Total rows processed\n- `records_inserted`: New records added\n- `records_updated`: Existing records updated\n- `errors`: List of errors (if any)",
        "operationId": "upload_stock_api_hospital_upload_stock_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_stock_api_hospital_upload_stock_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/hospital/upload-usage": {
      "post": {
        "tags": [
          "data upload"
        ],
        "summary": "Upload Usage",
        "description": "Upload historical usage data via CSV file (Hospital users only).\n\n**Authentication:** Requires hospital_admin or hospital_staff role\n\n**CSV Format:**\n```\nusage_date,medicine_id,medicine_name,quantity_consumed,department\n2024-01-15,MED001,Paracetamol 500mg,350,Outpatient\n```\n\n**Returns:**\n- `message`: Success/error message\n- `records_processed`: Total rows processed\n- `records_inserted`: New records added\n- `date_range`: Start and end dates of data\n- `errors`: List of errors (if any)",
        "operationId": "upload_usage_api_hospital_upload_usage_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_usage_api_hospital_upload_usage_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/hospital/procurement/recommendations": {
      "get": {
        "tags": [
          "procurement"
        ],
        "summary": "Get Procurement Recommendations",
        "description": "Get AI-recommended medicines to order\n\n**Returns:**\n- List of medicines that need reordering\n- Sorted by urgency (critical → high → medium)\n- Includes estimated costs and delivery times",
        "operationId": "get_procurement_recommendations_api_hospital_procurement_recommendations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/hospital/procurement/create-order": {
      "post": {
        "tags": [
          "procurement"
        ],
        "summary": "Create Procurement Order",
        "description": "Create procurement order from recommendations\n\n**Request Body:**\n- `medicines`: List of medicines to order with quantity and price\n- `expected_delivery_days`: Expected delivery timeline (default: 5)\n- `notes`: Optional order notes\n\n**Returns:**\n- List of created order IDs\n- Total cost and quantity\n- Order status (pending)",
        "operationId": "create_procurement_order_api_hospital_procurement_create_order_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_stock_api_hospital_upload_stock_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_stock_api_hospital_upload_stock_post"
      },
      "Body_upload_usage_api_hospital_upload_usage_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_usage_api_hospital_upload_usage_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrderCreate": {
        "properties": {
          "medicines": {
            "items": {
              "$ref": "#/components/schemas/OrderItemCreate"
            },
            "type": "array",
            "title": "Medicines"
          },
          "expected_delivery_days": {
            "type": "integer",
            "title": "Expected Delivery Days",
            "default": 5
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "medicines"
        ],
        "title": "OrderCreate"
      },
      "OrderCreateResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "order_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Order Ids"
          },
          "total_medicines": {
            "type": "integer",
            "title": "Total Medicines"
          },
          "total_quantity": {
            "type": "integer",
            "title": "Total Quantity"
          },
          "total_cost": {
            "type": "number",
            "title": "Total Cost"
          },
          "expected_delivery_date": {
            "type": "string",
            "format": "date",
            "title": "Expected Delivery Date"
          },
          "order_status": {
            "type": "string",
            "title": "Order Status"
          }
        },
        "type": "object",
        "required": [
          "message",
          "order_ids",
          "total_medicines",
          "total_quantity",
          "total_cost",
          "expected_delivery_date",
          "order_status"
        ],
        "title": "OrderCreateResponse"
      },
      "OrderItemCreate": {
        "properties": {
          "medicine_id": {
            "type": "string",
            "title": "Medicine Id"
          },
          "medicine_name": {
            "type": "string",
            "title": "Medicine Name"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "unit_price": {
            "type": "number",
            "title": "Unit Price"
          }
        },
        "type": "object",
        "required": [
          "medicine_id",
          "medicine_name",
          "quantity",
          "unit_price"
        ],
        "title": "OrderItemCreate"
      },
...
      "StockUploadResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "records_processed": {
            "type": "integer",
            "title": "Records Processed"
          },
          "records_inserted": {
            "type": "integer",
            "title": "Records Inserted"
          },
          "records_updated": {
            "type": "integer",
            "title": "Records Updated"
          },
          "errors": {
            "items": {

            },
            "type": "array",
            "title": "Errors",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "message",
          "records_processed",
          "records_inserted",
          "records_updated"
        ],
        "title": "StockUploadResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}